+++ Dockerfile.fe
@@ -6,4 +6,3 @@ COPY ./fe .
 RUN npm run build
 RUN npm install -g serve
 ENTRYPOINT ["npx", "serve", "-s", "dist"]
-
