+++ be/src/main/java/com/foodymoody/be/comment/controller/CommentController.java
@@ -3,20 +3,14 @@ package com.foodymoody.be.comment.controller;
 import com.foodymoody.be.comment.controller.dto.EditCommentRequest;
 import com.foodymoody.be.comment.controller.dto.RegisterCommentRequest;
 import com.foodymoody.be.comment.controller.dto.RegisterReplyRequest;
-import com.foodymoody.be.comment.controller.dto.ReplyResponse;
 import com.foodymoody.be.comment.domain.CommentId;
 import com.foodymoody.be.comment.service.CommentService;
 import com.foodymoody.be.common.annotation.MemberId;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.Sort.Direction;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -55,12 +49,5 @@ public class CommentController {
             @MemberId String memberId) {
         commentService.reply(id, request, memberId);
         return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @GetMapping("/api/comments/{id}/replies")
-    public ResponseEntity<Slice<ReplyResponse>> fetchAllReply(@PathVariable String id,
-            @PageableDefault(sort = "createdAt", direction = Direction.DESC) Pageable pageable) {
-        var allReplay = commentService.fetchAllReplay(id, pageable);
-        return ResponseEntity.ok(allReplay);
     }
 }
