+++ be/src/main/java/com/foodymoody/be/comment/service/CommentService.java
@@ -3,21 +3,17 @@ package com.foodymoody.be.comment.service;
 import com.foodymoody.be.comment.controller.dto.EditCommentRequest;
 import com.foodymoody.be.comment.controller.dto.RegisterCommentRequest;
 import com.foodymoody.be.comment.controller.dto.RegisterReplyRequest;
-import com.foodymoody.be.comment.controller.dto.ReplyResponse;
 import com.foodymoody.be.comment.domain.Comment;
 import com.foodymoody.be.comment.domain.CommentId;
 import com.foodymoody.be.comment.domain.Reply;
 import com.foodymoody.be.comment.domain.ReplyId;
 import com.foodymoody.be.comment.repository.CommentRepository;
-import com.foodymoody.be.comment.repository.ReplyRepository;
 import com.foodymoody.be.common.exception.CommentNotExistsException;
 import com.foodymoody.be.common.util.IdGenerator;
 import com.foodymoody.be.feed.service.FeedService;
 import java.time.LocalDateTime;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -27,7 +23,6 @@ public class CommentService {
 
     private final FeedService feedService;
     private final CommentRepository commentRepository;
-    private final ReplyRepository replyRepository;
     private final CommentMapper commentMapper;
 
     @Transactional
@@ -68,9 +63,5 @@ public class CommentService {
         comment.addReply(reply);
     }
 
-    @Transactional(readOnly = true)
-    public Slice<ReplyResponse> fetchAllReplay(String id, Pageable pageable) {
-        CommentId commentId = new CommentId(id);
-        return replyRepository.findReplyWithMemberAllById(commentId, pageable);
-    }
+
 }
